+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 - Fix depth write ignoring `DepthWriteMode` when returning a non-expression vertex record value
 - Avoid duplicate evaluation of complex vertex record expressions when writing output values
 - Remove `cs_` prefix for storage buffers declared as compute arguments
+- Allow inlining of mutable variables if not mutated in body
 
 ### 5.7.18
 - Add static class `Atomic` containing atomic integer operation intrinsics